refactor: Take buildHref out of kernel
Not a kernel function, merely an utilitarian one.
