+++ src/lib/index.test.ts
@@ -20,6 +20,7 @@ describe('index', () => {
             'isRouteActive',
             'activeBehavior',
             'Redirector',
+            'buildHref',
         ];
 
         // Act.
