+++ src/lib/kernel/index.test.ts
@@ -17,7 +17,6 @@ describe('index', () => {
             'LocationLite',
             'LocationFull',
             'preserveQueryInUrl',
-            'buildHref',
             'calculateMultiHashFragment'
         ];
